updated getFields and getMethods
Changed UMLClass's getFields() method to take a String fieldName parameter and return the corresponding DataFIeld rather than returning the hashmap which would need to be searched through by the caller. Similarly, changed getMethods() to require a String methodName and now it returns an ArrayList of overloaded methods rather than the class's entire hashmap of methods.