updated CommandRemove
made it so user can select from a list of overloaded list of methods to remove if they only specify the method name and not an accompanying list of parameters. Fixed bug where program would crash if user removed a relationship and then called List classes. Fixed a couple instances where there was no null checking after trying to return the ArrayList of methods associated with a methodName.