Updated guiclass so focus is removed with mouse clicks
Previously, the user had to press Enter after modifying a TextField for the changes to be accepted. Now they can also click anywhere else outside the TextField to remove focus and have the changes accepted. Also, modified the UMLMethod toString() so that it gives the properly formatted output for our TextFields.