Created screenshot functionality.
Added a promptScreenshotDirectory method to GUIFileBrowser which is used to find a location to save the screenshot. Wrote a screenshot command class responsible for actually capturing the screenshot.
Wrote a method in GuiController that responds to the export screenshot option being selected from the file drop down menu in the Gui.