Implemented Command Pattern
COmmand pattern consists of a client: exportScreenshotMenuAction() which is responsible for collecting data to create a concrete command: ScreenshotCommand class, which is responsible for implementing the execute() method of the ExportCommand interface. This ScreenshotCommand object is passed to a commandInvoker which blindly receives a command and calls its execute() function. This execute() function, in the case of a ScreenshotCommand object, will convert the world Pane into an image and store it in a user specified location.