Started implementing relevant classes
Created and finished the AStarNode class which represents a tile in the grid.
Created the PathGridMapper which will be responsible for conversions from JavaFX's continuous coordinates to the discrite coordinates of our tile grid. This class will also be responsible for establishing the initial perimeter of tiles around the source class box, and for having an isPassable() method to determine if the next prospective node in a path is contained in an obstacle or not.