Fixes some exceptions that got thrown
Couldn't get lines to draw during AStar due to some concurrentModificationExceptions. I've fixed those and now the line draws, but it isn't drawing an efficient line and It doesn't touch the source box, likely because the padding is too large for when we extract the perimeter nodes to populate the open set.